Connect authentication to frontend
Ensure that Vite server is always at port 8001 (for re-directing purposes). Back-end server is always going to be 8000 (default Django port).

Only show home page when user is logged in- so only people with accounts can use the app (to make this simple).

Currently, to (test) login go to http://localhost:8000/login with user: group50 and password: 123. Once logged in application should show the homepage at the redirected link http://localhost:8001/. If not logged in it displays an error message. To Logout, manually type in http://localhost:8000/logout for now (will add these links in frontend later).